remove FatsecretCore and migrate logic to fatsecret.py
